+++ se_invoice/src/Plugin/Block/MonthlyStatistics.php
@@ -14,7 +14,7 @@ use Drupal\se_report\ReportUtilityTrait;
  *   admin_label = @Translation("Invoice statistics per customer"),
  * )
  */
-class CustomerStatistics extends BlockBase {
+class MonthlyStatistics extends BlockBase {
 
   use ReportUtilityTrait;
 
@@ -32,7 +32,7 @@ class CustomerStatistics extends BlockBase {
       $month_data = [];
       $fg_colors = [];
       $bg_colors = [];
-      [$fg_color, $bg_color] = $this->generateColors();
+      [$fg_color, $bg_color] = $this->generateColorsDarkening(100, NULL, 50);
 
       foreach ($this->reportingMonths($year) as $month => $timestamps) {
         $query = \Drupal::entityQuery('node');
@@ -59,9 +59,13 @@ class CustomerStatistics extends BlockBase {
         'data' => $month_data,
         'backgroundColor' => $fg_colors,
         'borderColor' => $fg_colors,
-        'pointBackgroundColor' => $fg_colors,
-        'hoverBackgroundColor' => $bg_colors,
+        'hoverBackgroundColor' => $fg_colors,
         'fill' => FALSE,
+        'hover' => [
+          'mode' => 'dataset'
+        ],
+        'pointRadius' => 5,
+        'pointHoverRadius' => 10,
       ];
     }
 
@@ -72,9 +76,11 @@ class CustomerStatistics extends BlockBase {
       ],
       '#graph_type' => 'line',
       '#options' => [
-        'fill' => FALSE,
         'tooltips' => [
           'mode' => 'point'
+        ],
+        'hover' => [
+          'mode' => 'dataset'
         ],
       ],
       '#id' => 'invoice_statistics_customer',