+++ se_report/src/Plugin/Block/MonthlyStatistics.php
@@ -28,7 +28,7 @@ class MonthlyStatistics extends BlockBase {
       $month_data = [];
       $fg_colors = [];
       $bg_colors = [];
-      [$fg_color, $bg_color] = $this->generateColors();
+      [$fg_color, $bg_color] = $this->generateColorsDarkening(100, NULL, 50);
 
       foreach ($this->reportingMonths($year) as $month => $timestamps) {
         $query = $connection->select('node_field_data', 'nfd');
@@ -51,9 +51,13 @@ class MonthlyStatistics extends BlockBase {
         'data' => $month_data,
         'backgroundColor' => $fg_colors,
         'borderColor' => $fg_colors,
-        'pointBackgroundColor' => $fg_colors,
-        'hoverBackgroundColor' => $bg_colors,
+        'hoverBackgroundColor' => $fg_colors,
         'fill' => FALSE,
+        'hover' => [
+          'mode' => 'dataset'
+        ],
+        'pointRadius' => 5,
+        'pointHoverRadius' => 10,
       ];
     }
 
@@ -64,9 +68,11 @@ class MonthlyStatistics extends BlockBase {
       ],
       '#graph_type' => 'line',
       '#options' => [
-        'fill' => FALSE,
         'tooltips' => [
           'mode' => 'point'
+        ],
+        'hover' => [
+          'mode' => 'dataset'
         ],
       ],
       '#id' => 'monthly_statistics',